pyproject.toml
[tool.poetry]
name = "nmqn"
version = "0.1.0"
description = ""
readme = "README.md"
authors = ["takeshi <sci.and.eng@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
requests-html = "^0.9.0"
click = "^7.0"
pyppeteer = "^0.0.25"
pyyaml = "^3.13"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pytest-mock = "^1.10"
tox = "^3.6"

[tool.poetry.scripts]
nmqn = "nmqn.main:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
